AI-Models/Clustring_model/requirements.txt
joblib==1.4.2
mysql-connector-python==9.0.0
numpy==1.24.4
scikit-learn==1.0.2
scipy==1.10.1
threadpoolctl==3.5.0
